backend/requirements.txt
flask
flask-cors
pymongo
pillow
requests
python-dotenv
pytesseract
opencv-python
numpy
dotenv
